Restructuring Helpers (#167)
* Helper file split to format, derivative, controller, common

* helper.py deleted

* cleaned up some merge issues after testing
