+++ mailbagit/derivatives/eml.py
@@ -1,6 +1,7 @@
 # This is Eml derivative
 from os.path import join
-import mailbagit.helper as helper
+import mailbagit.helper.common as common
+import mailbagit.helper.derivative as derivative
 import os, glob
 import mailbox
 from mailbagit.email_account import EmailAccount
@@ -50,7 +51,7 @@ class EmlDerivative(Derivative):
             # Build msg
             if not message.Message and not message.Headers:
                 desc = "Unable to create EML as no body or headers present for " + str(message.Mailbag_Message_ID)
-                errors = helper.handle_error(errors, None, desc, "error")
+                errors = common.handle_error(errors, None, desc, "error")
             else:
                 fullObjectWrite = False
                 if message.Message:
@@ -73,9 +74,9 @@ class EmlDerivative(Derivative):
                                 outfile.close()
                             fullObjectWrite = True
                     except Exception as e:
-                        helper.deleteFile(filename + ".eml")
+                        derivative.deleteFile(filename + ".eml")
                         desc = "Error writing full email object to EML, generating it from the model instead"
-                        errors = helper.handle_error(errors, e, desc, "warn")
+                        errors = common.handle_error(errors, e, desc, "warn")
                         fullObjectWrite = False
 
                 if fullObjectWrite == False:
@@ -95,7 +96,7 @@ class EmlDerivative(Derivative):
                                 msg["X-Folder"] = message.Message_Path
                         except Exception as e:
                             desc = "Error writing headers for EML derivative"
-                            errors = helper.handle_error(errors, e, desc)
+                            errors = common.handle_error(errors, e, desc)
 
                         # Add message body
                         try:
@@ -109,10 +110,10 @@ class EmlDerivative(Derivative):
                                 msg.attach(alt)
                             else:
                                 desc = "No body present for " + str(message.Mailbag_Message_ID) + ". Created EML without message body"
-                                errors = helper.handle_error(errors, None, desc, "warn")
+                                errors = common.handle_error(errors, None, desc, "warn")
                         except Exception as e:
                             desc = "Error writing body for EML derivative"
-                            errors = helper.handle_error(errors, e, desc)
+                            errors = common.handle_error(errors, e, desc)
 
                         # Attachments
                         try:
@@ -121,7 +122,7 @@ class EmlDerivative(Derivative):
                                 if mimeType is None:
                                     mimeType = "text/plain"
                                     desc = "Mime type not found for attachment, set as " + mimeType
-                                    errors = helper.handle_error(errors, None, desc, "warn")
+                                    errors = common.handle_error(errors, None, desc, "warn")
                                 mimeType = mimeType.split("/")
                                 part = MIMEBase(mimeType[0], mimeType[1])
                                 part.set_payload(attachment.File)
@@ -130,7 +131,7 @@ class EmlDerivative(Derivative):
                                 msg.attach(part)
                         except Exception as e:
                             desc = "Error writing attachment(s) to EML derivative"
-                            errors = helper.handle_error(errors, e, desc)
+                            errors = common.handle_error(errors, e, desc)
 
                         # Write generated EML to disk
                         log.debug("Writing generated EML to " + str(out_dir))
@@ -144,15 +145,15 @@ class EmlDerivative(Derivative):
                                     outfile.close()
                         except Exception as e:
                             desc = "Error writing EML derivative"
-                            errors = helper.handle_error(errors, e, desc)
+                            errors = common.handle_error(errors, e, desc)
 
                     else:
                         desc = "Unable to create EML as no body or headers present for " + str(message.Mailbag_Message_ID)
-                        errors = helper.handle_error(errors, None, desc, "error")
+                        errors = common.handle_error(errors, None, desc, "error")
 
         except Exception as e:
             desc = "Error creating EML derivative"
-            errors = helper.handle_error(errors, e, desc)
+            errors = common.handle_error(errors, e, desc)
 
         message.Error.extend(errors["msg"])
         message.StackTrace.extend(errors["stack_trace"])
