+++ mailbagit/derivatives/pdf.py
@@ -3,7 +3,8 @@ import subprocess
 import distutils.spawn
 from mailbagit.derivative import Derivative
 from structlog import get_logger
-import mailbagit.helper as helper
+import mailbagit.helper.derivative as derivative
+import mailbagit.helper.common as common
 
 # only create format if pypff is successfully importable -
 # pst is not supported otherwise
@@ -62,15 +63,15 @@ if not skip_registry:
 
                 if message.HTML_Body is None and message.Text_Body is None:
                     desc = "No HTML or plain text body for " + str(message.Mailbag_Message_ID) + ", no PDF derivative created"
-                    errors = helper.handle_error(errors, None, desc, "warn")
+                    errors = common.handle_error(errors, None, desc, "warn")
                 else:
                     log.debug("Writing HTML to " + str(html_name) + " and converting to " + str(pdf_name))
                     # Calling helper function to get formatted html
                     try:
-                        html_formatted, encoding = helper.htmlFormatting(message, self.args.css)
+                        html_formatted, encoding = derivative.htmlFormatting(message, self.args.css)
                     except Exception as e:
                         desc = "Error formatting HTML for PDF derivative"
-                        errors = helper.handle_error(errors, e, desc)
+                        errors = common.handle_error(errors, e, desc)
 
                     if not self.args.dry_run:
                         try:
@@ -96,18 +97,18 @@ if not skip_registry:
                                     log.debug("Output converting to " + str(message.Mailbag_Message_ID) + ".pdf: " + stdout.decode("utf-8"))
                                 if stderr:
                                     desc = "Error converting to " + str(message.Mailbag_Message_ID) + ".pdf: " + stderr.decode("utf-8")
-                                    errors = helper.handle_error(errors, None, desc, "warn")
+                                    errors = common.handle_error(errors, None, desc, "warn")
                             # delete the HTML file
                             if os.path.isfile(pdf_name):
                                 os.remove(html_name)
 
                         except Exception as e:
                             desc = "Error writing HTML and converting to PDF derivative"
-                            errors = helper.handle_error(errors, e, desc)
+                            errors = common.handle_error(errors, e, desc)
 
             except Exception as e:
                 desc = "Error creating PDF derivative"
-                errors = helper.handle_error(errors, e, desc)
+                errors = common.handle_error(errors, e, desc)
 
             message.Error.extend(errors["msg"])
             message.StackTrace.extend(errors["stack_trace"])
