+++ mailbagit/derivatives/warc.py
@@ -1,7 +1,7 @@
 import os
 import warcio
-from structlog import get_logger
-import mailbagit.helper as helper
+import mailbagit.helper.derivative as derivative
+import mailbagit.helper.common as common
 from warcio.capture_http import capture_http
 from warcio import WARCWriter
 from warcio.statusandheaders import StatusAndHeaders
@@ -9,6 +9,8 @@ import requests  # requests *must* be imported after capture_http
 from threading import Thread
 import http.server
 import socketserver
+
+from structlog import get_logger
 
 log = get_logger()
 
@@ -39,7 +41,7 @@ class WarcDerivative(Derivative):
         if not self.args.dry_run:
             os.makedirs(self.warc_dir)
 
-            self.server_thread = Thread(target=helper.startServer, args=(self.args.dry_run, self.httpd, self.port))
+            self.server_thread = Thread(target=derivative.startServer, args=(self.args.dry_run, self.httpd, self.port))
             # Make it a daemon so it will stop after ctrl+c
             self.server_thread.daemon = True
             self.server_thread.start()
@@ -49,7 +51,7 @@ class WarcDerivative(Derivative):
         # Terminate the process
         try:
             if not self.args.dry_run:
-                helper.stopServer(self.args.dry_run, self.httpd[0])
+                derivative.stopServer(self.args.dry_run, self.httpd[0])
                 self.server_thread.join()
         except SystemExit:
             pass
@@ -71,7 +73,7 @@ class WarcDerivative(Derivative):
 
             if message.HTML_Body is None and message.Text_Body is None:
                 desc = "No HTML or plain text body for " + str(message.Mailbag_Message_ID) + ", no WARC derivative created"
-                errors = helper.handle_error(errors, None, desc, "warn")
+                errors = common.handle_error(errors, None, desc, "warn")
             else:
                 out_dir = os.path.join(self.warc_dir, message.Derivatives_Path)
                 filename = os.path.join(out_dir, str(message.Mailbag_Message_ID) + ".warc.gz")
@@ -84,13 +86,13 @@ class WarcDerivative(Derivative):
                     with open(filename, "wb") as output:
 
                         try:
-                            html_formatted, encoding = helper.htmlFormatting(message, self.args.css, headers=False)
+                            html_formatted, encoding = derivative.htmlFormatting(message, self.args.css, headers=False)
                             with open(self.tmp_file, "w", encoding="utf-8") as f:
                                 f.write(html_formatted)
                                 f.close()
                         except Exception as e:
                             desc = "Error formatting HTML for WARC derivative"
-                            errors = helper.handle_error(errors, e, desc)
+                            errors = common.handle_error(errors, e, desc)
 
                         writer = WARCWriter(output, gzip=True)
                         try:
@@ -103,7 +105,7 @@ class WarcDerivative(Derivative):
 
                         except requests.exceptions.HTTPError as e:
                             desc = "Error requesting HTML for WARC derivative"
-                            errors = helper.handle_error(errors, e, desc)
+                            errors = common.handle_error(errors, e, desc)
 
                         try:
                             # get raw headers from urllib3
@@ -119,20 +121,20 @@ class WarcDerivative(Derivative):
                             )
                         except Exception as e:
                             desc = "Error writing WARC headers"
-                            errors = helper.handle_error(errors, e, desc)
+                            errors = common.handle_error(errors, e, desc)
 
                         try:
                             writer.write_record(record)
                         except Exception as e:
                             desc = "Error writing WARC derivative"
-                            errors = helper.handle_error(errors, e, desc)
+                            errors = common.handle_error(errors, e, desc)
 
                     output.close()
-                    helper.deleteFile(self.tmp_file)
+                    derivative.deleteFile(self.tmp_file)
 
         except Exception as e:
             desc = "Error creating WARC derivative"
-            errors = helper.handle_error(errors, e, desc)
+            errors = common.handle_error(errors, e, desc)
 
         message.Error.extend(errors["msg"])
         message.StackTrace.extend(errors["stack_trace"])
